+++ activesupport/lib/active_support/descendants_tracker.rb
@@ -22,11 +22,18 @@ module ActiveSupport
 
       def clear
         if defined? ActiveSupport::Dependencies
+          # to_unload? is only defined in Zeitwerk mode.
+          to_unload = if Dependencies.respond_to?(:to_unload?)
+            ->(klass) { Dependencies.to_unload?(klass.name) }
+          else
+            ->(klass) { Dependencies.autoloaded?(klass) }
+          end
+
           @@direct_descendants.each do |klass, descendants|
-            if ActiveSupport::Dependencies.autoloaded?(klass)
+            if to_unload[klass]
               @@direct_descendants.delete(klass)
             else
-              descendants.reject! { |v| ActiveSupport::Dependencies.autoloaded?(v) }
+              descendants.reject! { |v| to_unload[v] }
             end
           end
         else