+++ activestorage/lib/active_storage/gem_version.rb
@@ -10,7 +10,7 @@ module ActiveStorage
     MAJOR = 6
     MINOR = 0
     TINY  = 0
-    PRE   = "beta3"
+    PRE   = "rc1"
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
   end